Created Components & added HomeScreen component & StackNavigator component
