+++ App.js
@@ -1,25 +1,14 @@
 import React from 'react';
 
-import {View, Text, StyleSheet} from 'react-native';
+import {NavigationContainer} from '@react-navigation/native';
+import StackScreen from './Components/Navigators/StackNavigator';
 
 const App = () => {
   return (
-    <View style={styles.container}>
-      <Text style={styles.text}>Hello world</Text>
-    </View>
+    <NavigationContainer>
+      <StackScreen />
+    </NavigationContainer>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  text: {
-    fontSize: 25,
-    fontWeight: 'bold',
-  },
-});
 
 export default App;
