+++ Components/HomeScreen.js
@@ -1,0 +1,25 @@
+import React from 'react';
+
+import {View, Text, StyleSheet} from 'react-native';
+
+const HomeScreen = () => {
+  return (
+    <View style={styles.container}>
+      <Text style={styles.text}>Hello world</Text>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  text: {
+    fontSize: 25,
+    fontWeight: 'bold',
+  },
+});
+
+export default HomeScreen;
